requirements.txt
Django==5.2.5
Pillow==11.3.0
requests==2.32.4

# Production Dependencies
psycopg2-binary==2.9.9  # PostgreSQL adapter
gunicorn==21.2.0  # WSGI server
django-environ==0.11.2  # Environment variables
whitenoise==6.6.0  # Static files serving
dj-database-url==2.1.0  # Database URL parsing

# WebSocket Support
channels==4.0.0
daphne==4.0.0

# Media Storage
django-storages==1.14.2  # AWS S3 support
boto3==1.34.0  # AWS SDK

# Alternative: Cloudinary
cloudinary==1.36.0

# Security & Monitoring  
django-ratelimit==4.1.0  # Rate limiting
django-cors-headers==4.3.1  # CORS support
sentry-sdk==1.39.2  # Error tracking

# Performance
redis==5.0.1  # Caching
django-redis==5.4.0  # Redis cache backend
channels-redis==4.2.0  # WebSocket with Redis